fastapi-microservices/auth-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.33.0
python-multipart==0.0.20
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.27.2
